[TNT-140] move: data 모듈 -> data:network 모듈로 세분화
