+++ data/network/src/main/AndroidManifest.xml
