+++ data/network/src/main/java/co/kr/data/network/repository/TnTRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package co.kr.tnt.data.repository
+package co.kr.data.network.repository
 
 import co.kr.tnt.domain.repository.TnTRepository
 import javax.inject.Inject
