+++ data/network/src/main/java/co/kr/data/network/service/TnTService.kt
@@ -1,0 +1,3 @@
+package co.kr.data.network.service
+
+internal interface TnTService
