+++ data/network/src/main/java/co/kr/data/network/tnt/response/TnTResponse.kt
@@ -1,15 +1,15 @@
-package co.kr.tnt.data.model.tnt.request
+package co.kr.data.network.tnt.response
 
 import co.kr.tnt.domain.model.TnT
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class TnTRequest(
+data class TnTResponse(
     @SerialName("id")
     val id: String,
 )
 
-fun TnT.toRemote(): TnTRequest = TnTRequest(
-    id = this.id,
+fun TnTResponse.toDomain(): TnT = TnT(
+    id = id,
 )
