--- data/src/main/java/co/kr/tnt/data/service/TnTService.kt
@@ -1,3 +1,0 @@
-package co.kr.tnt.data.service
-
-internal interface TnTService
